Adds ErrorWrappingMiddleware
This will effectively intercept the instance where we would provide an invalid route to Mvc. In the middleware, context.Response.StatusCode is 404 for this case. We do not know at this time what other status codes this middleware would intercept. At this point, we need to set this up before we proceed to the next task - creating a uniform Api response object.