Adds consistent ApiResponse data format
All responses back to the user will now have a consistent object to return, whether it's a 400, 404, 200, 201, or 204. We have implemented this in the Values controller and in our ErrorWrappingMiddleware. The curious case of 401 issued by ASP.net Core when the request butts up against violation of the Authorize attribute's roles or policies, or when we actually get an exception, has not yet been examined.