Able to catch Authorize attr violations in ResourceServerMiddleware
While this surely catches the instance where a request violates the Authorize decorator on an API function (e.g. no authorization bearer header, or auth bearer with token included but insufficient credentials for that Authorize decorator), we don't get the exact reason why the request was denied. For now, we will merely return an UnauthorizeApiResponse with a generic message of "insufficient-credentials"